Change img file names.
People who have cloned my repository couldn't run the app on localhost because of an issue with the old file extension names. Renaming the file extension to .png in all lowecase wasn't actually working to commit that to github. Instead renamed the file names completely and took the oppurtunity to rewrite the file extensions.